fix: 컴포넌트를 여러 곳에서 다른 인자로 전달받아 오류 발생
